Move config gen modules out of fedimint-server
