+++ fedimint-server/Cargo.toml
@@ -27,7 +27,6 @@ jsonrpsee = { version = "0.16.2", features = ["server"] }
 mint-client = { path = "../client/client-lib" }
 fedimint-api = { path = "../fedimint-api" }
 fedimint-core = { path = "../fedimint-core" }
-fedimint-wallet = { path = "../modules/fedimint-wallet", features = ["native"] }
 rand = "0.8"
 rayon = "1.6.1"
 rcgen = "=0.10.0"
